[FEAT] 랜덤으로 상대방(컴퓨터)의 수를 설정하는 기능 구현
