+++ docs/README.md
@@ -2,17 +2,17 @@
 
 ## 🔍 기능 목록
 
-- 게임을 시작하는 기능
-- 랜덤으로 상대방(컴퓨터)의 수를 설정하는 기능
-- 사용자로부터 값을 입력받는 기능
-  - 사용자로부터 값을 받았을 때 예외 처리
-  - 예외 일때 알려주는 기능
-- 스트라이크, 볼, 낫싱으로 결과를 내는 기능
-- 결과를 출력하는 기능
-- 3스트라이크면
-  - 3스트라이크면 게임을 종료하는 기능
-  - 다시 시작하는 기능
-- 예외 처리
+- [ ]게임을 시작하는 기능
+- [x]랜덤으로 상대방(컴퓨터)의 수를 설정하는 기능
+- [ ]사용자로부터 값을 입력받는 기능
+  - [ ]사용자로부터 값을 받았을 때 예외 처리
+  - [ ]예외 일때 알려주는 기능
+- [ ]스트라이크, 볼, 낫싱으로 결과를 내는 기능
+- [ ]결과를 출력하는 기능
+- [ ]3스트라이크면
+  - [ ]3스트라이크면 게임을 종료하는 기능
+  - [ ]다시 시작하는 기능
+- [ ]예외 처리
 
 ## 🔍 유의 사항
