+++ src/components/InputNumber.js
@@ -1,0 +1,3 @@
+import { MissionUtils } from "@woowacourse/mission-utils";
+
+export default function InputNumber() {}
